+++ doc/sphinx/changelog.rst
@@ -7,6 +7,7 @@ This release series is targeted to DAS production stability and quality.
 
 - 1.1.X
 
+  - Retire DASLogger in favor of new PrintManager
   - code re-factoring to address duplicates issue; ticket #2848
   - add dataset/block/replica presence, according to ticket #2858; made changes to maps